+++ ci/test.sh
@@ -10,3 +10,6 @@ mix deps.get
 mix compile
 mix credo --strict
 mix test
+mix sobelow --private --compact --ignore Config.Secrets,Config.HTTPS --router lib/td_cx_web/router.ex
+mix sobelow --private --compact --ignore Config.Secrets,Config.HTTPS --router lib/td_dd_web/router.ex
+mix sobelow --private --compact --ignore Config.Secrets,Config.HTTPS --router lib/td_dq_web/router.ex