+++ lib/td_dd/data_structures/audit.ex
@@ -9,7 +9,6 @@ defmodule TdDd.DataStructures.Audit do
 
   alias Ecto.Changeset
   alias TdCache.TaxonomyCache
-  alias TdDd.Repo
 
   @doc """
   Publishes a `:structure_note_updated` event when modifying a StructureNote. Should be called using `Ecto.Multi.run/5`.
@@ -221,34 +220,6 @@ defmodule TdDd.DataStructures.Audit do
   end
 
   @doc """
-  Publishes a `:grant_approvals` event when creating a Grant. Should be called using `Ecto.Multi.run/5`.
-  """
-
-  def grant_request_approval_created(_repo, %{
-        approval: %{grant_request_id: id, user_id: user_id, comment: comment} = approval,
-        status: %{status: "rejected" = status}
-      }) do
-    %{
-      grant_request: %{
-        group: %{user_id: recipient_id},
-        data_structure: %{current_version: %{name: data_structure_name}}
-      }
-    } = Repo.preload(approval, grant_request: [:group, data_structure: :current_version])
-
-    payload = %{
-      name: data_structure_name,
-      status: status,
-      recipient_ids: [recipient_id],
-      comment: comment,
-      user_id: user_id
-    }
-
-    publish("grant_approval", "grant_requests", id, user_id, payload)
-  end
-
-  def grant_request_approval_created(_repo, _), do: {:ok, nil}
-
-  @doc """
   Publishes a `:grant_created` event when creating a Grant. Should be called using `Ecto.Multi.run/5`.
   """
   def grant_created(_repo, %{grant: %{id: id} = grant, latest: latest}, user_id) do
@@ -298,6 +269,10 @@ defmodule TdDd.DataStructures.Audit do
 
   defp with_domain_ids(%Changeset{} = changeset, %{data_structure: %{domain_ids: domain_ids}}) do
     Changeset.put_change(changeset, :domain_ids, get_domain_ids(domain_ids))
+  end
+
+  defp with_domain_ids(%{domain_ids: acc_domain_ids} = payload, %{data_structure: %{domain_ids: domain_ids}}) do
+    Map.put(payload, :domain_ids, acc_domain_ids ++ get_domain_ids(domain_ids))
   end
 
   defp with_domain_ids(%{} = payload, %{data_structure: %{domain_ids: domain_ids}}) do