+++ lib/td_dd_web/controllers/data_structure_version_controller.ex
@@ -1,6 +1,5 @@
 defmodule TdDdWeb.DataStructureVersionController do
   use TdDdWeb, :controller
-  use TdHypermedia, :controller
   use PhoenixSwagger
 
   import Canada, only: [can?: 2]
@@ -28,6 +27,7 @@ defmodule TdDdWeb.DataStructureVersionController do
     :domain,
     :external_id,
     :links,
+    :with_protected_metadata,
     :metadata_versions,
     :parents,
     :profile,
@@ -78,10 +78,20 @@ defmodule TdDdWeb.DataStructureVersionController do
 
   defp enrich_opts(%{user_id: user_id} = claims, data_structure) do
     Enum.filter(@enrich_attrs, fn
-      :profile -> can?(claims, view_data_structures_profile(data_structure))
-      :with_confidential -> can?(claims, manage_confidential_structures(data_structure))
-      :grants -> can?(claims, view_grants(data_structure))
-      _ -> true
+      :profile ->
+        can?(claims, view_data_structures_profile(data_structure))
+
+      :with_confidential ->
+        can?(claims, manage_confidential_structures(data_structure))
+
+      :grants ->
+        can?(claims, view_grants(data_structure))
+
+      :with_protected_metadata ->
+        can?(claims, view_protected_metadata(data_structure))
+
+      _ ->
+        true
     end) ++ [user_id: user_id]
   end
 
@@ -101,26 +111,24 @@ defmodule TdDdWeb.DataStructureVersionController do
         view_profiling_permission: can?(claims, view_data_structures_profile(data_structure)),
         profile_permission: can?(claims, profile(dsv)),
         request_grant: can_request_grant?(claims, data_structure),
-        update_grant_removal: can_update_grant_removal?(claims, data_structure)
+        update_grant_removal: can_update_grant_removal?(claims, data_structure),
+        create_foreign_grant_request: can?(claims, create_foreign_grant_request(data_structure))
       }
 
       render(conn, "show.json",
         data_structure_version: dsv,
         tags: tags,
         user_permissions: user_permissions,
-        actions: actions(claims, data_structure),
-        hypermedia: hypermedia("data_structure_version", conn, dsv)
+        actions: actions(claims, dsv)
       )
     else
       render_error(conn, :forbidden)
     end
   end
 
-  defp actions(claims, data_structure) do
-    if can?(claims, link_data_structure_tag(data_structure)) do
-      %{manage_tags: Tags.list_available_tags(data_structure)}
-    else
-      %{}
+  defp actions(claims, dsv) do
+    if can?(claims, create_link(dsv)) do
+      %{create_link: true}
     end
   end
 
@@ -138,7 +146,10 @@ defmodule TdDdWeb.DataStructureVersionController do
   end
 
   defp get_data_structure_version(data_structure_id, "latest", opts) do
-    DataStructures.get_latest_version(data_structure_id, opts)
+    DataStructures.get_latest_version(
+      data_structure_id,
+      opts
+    )
   end
 
   defp get_data_structure_version(data_structure_id, version, opts) do