+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - [TD-5076] Allow to clone implementations for non admin users
 - [TD-5169] Improve list_data_structure_versions since query performance
 - [TD-5170] Capture graph tests Logger.warn output
+- [TD-4214] Publish grant requests workflow events
 
 ## [4.50.0] 2022-09-05